Converted CompanyProfile & CreateCompany into Fragments. Re-designed Navigation Drawer to fit with Material3 Standards. Added Bottom Nav & custom App Bar. Users are no longer forced to create/join a company upon launching the app if they don't belong to one yet. Fixed issues related to saving current user preferences with Google Sign-On by implementing CoroutineSuspend. DataRepository Functions that require CompanyId as parameter now properly checks for null companyId values to avoid firestore exception error.

+++ app/src/main/java/com/naozumi/izinboss/view/UserProfileFragment.kt
@@ -51,9 +51,18 @@ class UserProfileFragment : Fragment() {
         }
 
         binding?.btnDeleteAccount?.setOnClickListener(2000L) {
-            lifecycleScope.launch {
-                deleteUserData(user?.uid)
-                viewModel.deleteCurrentUserDataStore()
+            AlertDialog.Builder(requireActivity()).apply {
+                setTitle("Warning")
+                setMessage("Are you sure you want to delete your account?")
+                setPositiveButton("Yes") { _, _ ->
+                    lifecycleScope.launch {
+                        deleteUserData(user?.uid)
+                        viewModel.deleteCurrentUserDataStore()
+                    }
+                }
+                setNegativeButton("No") { _, _ -> }
+                create()
+                show()
             }
         }
     }
